Moved code into __init__.py
